+++ src/predicates/Shape.ts
@@ -29,7 +29,7 @@ export default class ShapePredicate<T extends object> extends Predicate<T> {
     Object.keys(this.contents).forEach(baseKey => {
       const key = baseKey as keyof T;
 
-      struct[key] = this.contents[key].default();
+      struct[key] = this.contents[key].default()!;
     });
 
     return struct as T;